Add files via upload
There are many different forms of heart disease. The most common cause of heart disease is narrowing or blockage of the coronary arteries, the blood vessels that supply blood to the heart itself. This is called coronary artery disease and happens slowly over time. It's the major reason people have heart attacks.
Coronary artery disease, arrhythmia, heart valve disease and heart failure are the four most common types of heart disease. Heart disease refers to several types of conditions that affect the heart. Heart disease is the leading cause of death in the U.S., causing about 1 in 4 deaths.